+++ showcase/main.ts
@@ -1,0 +1,6 @@
+import { createApp } from 'vue'
+
+import KitShowcase from './KitShowcase.vue'
+import './main.css'
+
+createApp(KitShowcase).mount('#app')